Use UInt64 for token counts
Otherwise, it wraps at ~2B, although this one wraps at ~8B (weird): https://wandb.ai/incite-mist/mist/runs/fr1h7izc?nw=nwuserawadell

Anywho uint64, should would up to ~1.8e19 tokens

+++ electrolyte_fm/utils/metrics.py
@@ -243,10 +243,10 @@ class TokenCounter(Metric):
     def __init__(self) -> None:
         super().__init__()
         self.add_state(
-            "masked_tokens", torch.tensor(0, dtype=torch.int), dist_reduce_fx="sum"
+            "masked_tokens", torch.tensor(0, dtype=torch.uint64), dist_reduce_fx="sum"
         )
         self.add_state(
-            "total_tokens", torch.tensor(0, dtype=torch.int), dist_reduce_fx="sum"
+            "total_tokens", torch.tensor(0, dtype=torch.uint64), dist_reduce_fx="sum"
         )
 
     def update(
